+++ packages/lumx-react/src/components/radio-button/RadioButton.tsx
@@ -1,4 +1,4 @@
-import React, { useMemo, forwardRef, ReactNode, SyntheticEvent } from 'react';
+import React, { useMemo, forwardRef, ReactNode, SyntheticEvent, InputHTMLAttributes } from 'react';
 
 import classNames from 'classnames';
 import { uid } from 'uid';
@@ -73,7 +73,7 @@ export const RadioButton: Comp<RadioButtonProps, HTMLDivElement> = forwardRef((p
         value,
         ...forwardedProps
     } = props;
-    const radioButtonId = useMemo(() => id || `${CLASSNAME.toLowerCase()}-${uid()}`, [id]);
+    const inputId = useMemo(() => id || `${CLASSNAME.toLowerCase()}-${uid()}`, [id]);
 
     const handleChange = (event: React.ChangeEvent<HTMLInputElement>) => {
         if (onChange) {
@@ -101,13 +101,14 @@ export const RadioButton: Comp<RadioButtonProps, HTMLDivElement> = forwardRef((p
                     ref={inputRef}
                     className={`${CLASSNAME}__input-native`}
                     disabled={isDisabled}
-                    id={radioButtonId}
+                    id={inputId}
                     tabIndex={isDisabled ? -1 : 0}
                     type="radio"
                     name={name}
                     value={value}
                     checked={isChecked}
                     onChange={handleChange}
+                    aria-describedby={helper ? `${inputId}-helper` : undefined}
                 />
 
                 <div className={`${CLASSNAME}__input-placeholder`}>
@@ -118,12 +119,12 @@ export const RadioButton: Comp<RadioButtonProps, HTMLDivElement> = forwardRef((p
 
             <div className={`${CLASSNAME}__content`}>
                 {label && (
-                    <InputLabel htmlFor={radioButtonId} theme={theme} className={`${CLASSNAME}__label`}>
+                    <InputLabel htmlFor={inputId} theme={theme} className={`${CLASSNAME}__label`}>
                         {label}
                     </InputLabel>
                 )}
                 {helper && (
-                    <InputHelper theme={theme} className={`${CLASSNAME}__helper`}>
+                    <InputHelper id={`${inputId}-helper`} theme={theme} className={`${CLASSNAME}__helper`}>
                         {helper}
                     </InputHelper>
                 )}