+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Fixed
+
+-   RadioButton: fix helper text accessibility
+
 ## [3.0.4][] - 2022-11-07
 
 ### Added